Added job to delete content of database and added fresh new data for that, refactored controllers and tests, added rate limiting, modifying Xunit tests

+++ LibraryAppWebAPI/Controllers/DvdsController.cs
@@ -5,6 +5,7 @@ using Swashbuckle.AspNetCore.Annotations;
 using LibraryAppWebAPI.Models;
 using LibraryAppWebAPI.Models.DTOs;
 using LibraryAppWebAPI.Repository.Interfaces;
+using LibraryAppWebAPI.Models.RequestModels;
 
 namespace LibraryAppWebAPI.Controllers;
 
@@ -13,18 +14,28 @@ namespace LibraryAppWebAPI.Controllers;
 [SwaggerTag("Dvds")]
 public class DvdsController(IDvdRepository dvdRepository, IRentalEntryRepository rentalEntryRepository) : ControllerBase
 {
+    private static readonly Dictionary<string, DateTime> LastRequestTimes = new();
+    private static readonly object LockObject = new();
+
     // GET: api/Dvds
     [HttpGet]
     [ProducesResponseType(200, Type = typeof(OkResult))]
     [ProducesResponseType(404, Type = typeof(NotFound))]
     [SwaggerOperation(Summary = "Get all dvds", Tags = ["Dvds"])]
-    public ActionResult<IEnumerable<Dvd>> GetDvds()
+    public ActionResult<IEnumerable<DvdRequestModel>> GetDvds()
     {
         IEnumerable<Dvd> dvds = dvdRepository.GetAll();
         if (dvds == null || !dvds.Any())
             return NotFound("No dvds in database");
+
+        List<DvdRequestModel> dvdRequestModelList = new List<DvdRequestModel>();
+        foreach (Dvd dvd in dvds)
+        {
+            DvdRequestModel dvdRequestModel = new (dvd);
+            dvdRequestModelList.Add(dvdRequestModel);
+        }
         
-        return Ok(dvds);
+        return Ok(dvdRequestModelList);
     }
 
     // GET: api/Dvds/5
@@ -32,14 +43,16 @@ public class DvdsController(IDvdRepository dvdRepository, IRentalEntryRepository
     [ProducesResponseType(200, Type = typeof(OkResult))]
     [ProducesResponseType(404, Type = typeof(NotFound))]
     [SwaggerOperation(Summary = "Get dvd by id", Tags = ["Dvds"])]
-    public ActionResult<Dvd> GetDvd(int id)
+    public ActionResult<DvdRequestModel> GetDvd(int id)
     {
         Dvd dvd = dvdRepository.GetById(id);
 
         if (!dvdRepository.DvdExists(id))
             return NotFound($"Dvd with id {id} does not exist");
 
-        return Ok(dvd);
+        DvdRequestModel dvdRequestModel = new (dvd);
+
+        return Ok(dvdRequestModel);
     }
 
     // POST: api/Dvds
@@ -52,6 +65,21 @@ public class DvdsController(IDvdRepository dvdRepository, IRentalEntryRepository
         if (!ModelState.IsValid)
             return BadRequest(ModelState);
 
+        string clientIp = HttpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
+
+        lock (LockObject)
+        {
+            if (LastRequestTimes.TryGetValue(clientIp, out DateTime lastRequestTime))
+            {
+                if ((DateTime.UtcNow - lastRequestTime).TotalSeconds < 10) // Limit: 10 sekúnd medzi požiadavkami
+                {
+                    return StatusCode(429, "You are sending requests too quickly.");
+                }
+            }
+
+            LastRequestTimes[clientIp] = DateTime.UtcNow;
+        }
+
         Dvd dvd = new();
         {
             dvd.Author = dvdRequest.Author;
@@ -60,10 +88,11 @@ public class DvdsController(IDvdRepository dvdRepository, IRentalEntryRepository
             dvd.TotalAvailableCopies = dvdRequest.TotalAvailableCopies;
             dvd.PublishYear = dvdRequest.PublishYear;
             dvd.NumberOfMinutes = dvdRequest.NumberOfMinutes;
+            dvd.CanManipulate = true;
         }
 
         dvdRepository.Create(dvd);
-        return CreatedAtAction("GetDvd", new { id = dvd.Id }, dvd);
+        return CreatedAtAction("GetDvd", new { id = dvd.Id }, dvdRequest);
     }
 
     // PUT: api/Dvds/5
@@ -80,6 +109,24 @@ public class DvdsController(IDvdRepository dvdRepository, IRentalEntryRepository
         if (!dvdRepository.DvdExists(id))
             return NotFound($"Dvd with id {id} does not exist");
 
+        string clientIp = HttpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
+
+        lock (LockObject)
+        {
+            if (LastRequestTimes.TryGetValue(clientIp, out DateTime lastRequestTime))
+            {
+                if ((DateTime.UtcNow - lastRequestTime).TotalSeconds < 10) // Limit: 10 sekúnd medzi požiadavkami
+                {
+                    return StatusCode(429, "You are sending requests too quickly.");
+                }
+            }
+
+            LastRequestTimes[clientIp] = DateTime.UtcNow;
+        }
+
+        if (!dvdRepository.CanManipulate(id))
+            return BadRequest($"You can't update this DVD!");
+
         Dvd dvd = dvdRepository.GetById(id);
         {
             dvd.Author = dvdRequest.Author;
@@ -88,6 +135,7 @@ public class DvdsController(IDvdRepository dvdRepository, IRentalEntryRepository
             dvd.TotalAvailableCopies = dvdRequest.TotalAvailableCopies;
             dvd.PublishYear = dvdRequest.PublishYear;
             dvd.NumberOfMinutes = dvdRequest.NumberOfMinutes;
+            dvd.CanManipulate = true;
         }
 
         if (!rentalEntryRepository.RentalEntryByTitleIdExist(id))
@@ -112,14 +160,13 @@ public class DvdsController(IDvdRepository dvdRepository, IRentalEntryRepository
         if (!dvdRepository.DvdExists(id))
             return NotFound($"Dvd with id {id} does not exist");
 
+        if (!dvdRepository.CanManipulate(id))
+            return BadRequest($"You can't delete this DVD!");
+
         if (!rentalEntryRepository.RentalEntryByTitleIdExist(id))
-        {
             dvdRepository.Delete(id);
-        }
         else
-        {
             return BadRequest($"This title was found in rentals. This title cannot be removed");
-        }
 
         return Ok($"Dvd with id {id} was successfully deleted");
     }
